requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
beanie>=1.11.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator>=1.1.3
fastapi>=0.68.0
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor>=2.5.0
orjson==3.10.15
passlib>=1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic>=1.8.0
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv>=0.19.0
python-jose>=3.3.0
python-multipart>=0.0.5
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
toml==0.10.2
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
uvicorn>=0.15.0
watchfiles==1.0.4
websockets==15.0.1